requirements.txt
agentlightning
pydantic
pyyaml
pandas
numpy
yfinance
rich
textblob
matplotlib
